(BSR)[PRO] refactor: yet remove another useEffect
