+++ pro/src/components/OfferIndividualForm/hooks/index.ts
@@ -1,2 +1,1 @@
-export { useFilteredVenueList } from './useFilteredVenueList'
 export { useShowFullForm } from './useShowFullForm'
