--- pro/src/components/OfferIndividualForm/hooks/useFilteredVenueList/index.ts
@@ -1,1 +1,0 @@
-export { default as useFilteredVenueList } from './useFilteredVenueList'
