+++ pro/src/components/OfferIndividualForm/utils/getFilteredVenueList.ts
@@ -1,0 +1,33 @@
+import { CATEGORY_STATUS } from 'core/Offers'
+import { IOfferSubCategory } from 'core/Offers/types'
+import { TOfferIndividualVenue } from 'core/Venue/types'
+
+export interface IUseFilteredVenueListProps {
+  subCategories: IOfferSubCategory[]
+  subcategoryId: string
+  venueList: TOfferIndividualVenue[]
+}
+
+export const getFilteredVenueList = (
+  subcategoryId: string,
+  subCategories: IOfferSubCategory[],
+  venueList: TOfferIndividualVenue[]
+): TOfferIndividualVenue[] => {
+  if (!subcategoryId) {
+    return venueList
+  }
+
+  const subCategory = subCategories.find(s => s.id === subcategoryId)
+  if (
+    !subCategory ||
+    subCategory.onlineOfflinePlatform === CATEGORY_STATUS.ONLINE_OR_OFFLINE
+  ) {
+    return venueList
+  }
+
+  return venueList.filter(venue =>
+    subCategory.onlineOfflinePlatform === CATEGORY_STATUS.ONLINE
+      ? venue.isVirtual
+      : !venue.isVirtual
+  )
+}
